Module cryptographic_message_syntax::asn1::rfc5652

ASN.1 data structures defined by RFC 5652.

The types defined in this module are intended to be extremely low-level and only to be used for (de)serialization. See types outside the asn1 module tree for higher-level functionality.

Some RFC 5652 types are defined in the x509-certificate crate, which this crate relies on for certificate parsing functionality.

Structs

AuthenticatedData

Authenticated data.

CertificateSet

ContentInfo

Content info.

DigestAlgorithmIdentifiers

Digest algorithm identifiers.

DigestedData

Digested data.

EncapsulatedContentInfo

Encapsulated content info.

EncryptedContentInfo

Encrypted content info.

EncryptedData

Encrypted data.

EnvelopedData

Enveloped-data content type.

IssuerAndSerialNumber

Issuer and serial number.

KekIdentifier

Key encryption key identifier.

KekRecipientInfo

Key encryption key recipient info.

KeyAgreeRecipientInfo

Key agreement recipient info.

KeyTransRecipientInfo

Key trans recipient info.

OriginatorInfo

Originator info.

OriginatorPublicKey

Originator public key.

OtherCertificateFormat

Other certificate format.

OtherKeyAttribute

Other key attribute.

OtherRecipientInfo

OtherRevocationInfoFormat

Other revocation info format.

PasswordRecipientInfo

Password recipient info.

RecipientEncryptedKey

Recipient encrypted key.

RecipientKeyIdentifier

Recipient key identifier.

RevocationInfoChoices

Revocation info choices.

SignedAttributes

Signed attributes.

SignedAttributesDer

SignedData

Represents signed data.

SignerInfo

Per-signer information.

SignerInfos

Signer infos.

UnsignedAttributes

Unsigned attributes.

Enums

CertificateChoices

Certificate choices.

CmsVersion

Version number.

KeyAgreeRecipientIdentifier

Key agreement recipient identifier.

OriginatorIdentifierOrKey

Originator identifier or key.

RecipientIdentifier

Recipient identifier.

RecipientInfo

Recipient info.

RevocationInfoChoice

Revocation info choice.

SignerIdentifier

Identifies the signer.

Time

Time variant.

Constants

OID_AUTHENTICATED_DATA

Authenticated-data content type.

OID_CONTENT_TYPE

Identifies the content-type attribute.

OID_COUNTER_SIGNATURE

Identifies the countersignature attribute.

OID_DIGESTED_DATA

Digested-data content type.

OID_ENCRYPTED_DATA

Encrypted-data content type.

OID_ENVELOPE_DATA

Enveloped data content type.

OID_ID_DATA

The data content type.

OID_ID_SIGNED_DATA

The signed-data content type.

OID_MESSAGE_DIGEST

Identifies the message-digest attribute.

OID_SIGNING_TIME

Identifies the signing-time attribute.

Type Definitions

AttributeCertificateV2

AuthAttributes

CertificateSerialNumber

ContentEncryptionAlgorithmIdentifier

ContentType

CounterSignature

Digest

DigestAlgorithmIdentifier

EncryptedContent

EncryptedKey

KeyDerivationAlgorithmIdentifier

KeyEncryptionAlgorithmIdentifier

MessageAuthenticationCode

MessageAuthenticationCodeAlgorithm

MessageDigest

RecipientInfos

SignatureAlgorithmIdentifier

SignatureValue

SigningTime

UnauthAttributes

UnprotectedAttributes

UserKeyingMaterial